Fixed external project compilation: libconfig.
Libconfig 1.4.8 comes with a set of unit tests that depend on the
tinytest library. The libconfig dependency on tinytest is phrased in
terms of relative paths, hence out-of-source compilations may fail.
Our external projects adds a patch that disables building tests and
tinytest. The tests should be re-enabled once the libconfig team fiexes
the issue.

Signed-off-by: Marc Kirchner <mail@marc-kirchner.de>